add ui and logic for grouping
